@@ -5,7 +5,7 @@ author: Carlos E. Torres <cetorres@cetorres.com>
 homepage: https://github.com/cetorres/dropdown_formfield
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
